reorganize context folder
