+++ client/src/components/input/sortSelector/sortSelector.test.jsx
@@ -1,6 +1,6 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from "@testing-library/user-event";
-import { SortProvider } from '../../contexts/sort';
+import { SortProvider } from '../../../contexts/sort';
 import SortSelector from './sortSelector';
 
 const MockWrapper = ({ children }) => (
