+++ client/src/components/jobsPage/jobListing/jobListing.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useState } from 'react';
 import { Dialog, Icon } from '@blueprintjs/core';
 
-import { useUserContext } from '../../contexts/user';
+import { useUserContext } from '../../../contexts/user';
 import { numberFormatter } from '../../utils';
 import CITIES_UNTYPED from '../../../assets/cities.json';
 import JobDetails from './jobDetails/jobDetails';
