+++ client/src/components/jobsPage/sidebar/utils/sortDefined.tsx
@@ -5,7 +5,7 @@ import { BlueprintIcons_16Id } from '@blueprintjs/icons/lib/esm/generated/16px/b
 import Sorts from '../../../input/sortSelector/sortSelector';
 import PrimaryButton from '../../../buttons/primaryButton';
 import SecondaryButton from '../../../buttons/secondaryButton';
-import css from '../../../contexts/themes.scss';
+import css from '../../../../contexts/themes.scss';
 
 type sortDefinedProps = {
   sortOrder: string,
