+++ client/src/components/welcome/welcome.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { NavigateFunction, useNavigate } from 'react-router-dom';
 
-import { useUserContext } from '../contexts/user';
+import { useUserContext } from '../../contexts/user';
 import {
   footer,
   headerAndLogo,
