+++ client/src/contexts/filteredJobs.tsx
@@ -1,6 +1,6 @@
 // Package imports
 import React, { createContext,Dispatch, SetStateAction, useContext, useState } from 'react';
-import { job, Children } from '../../interfaces'
+import { job, Children } from '../interfaces'
 
 const Context = createContext<[job[], Dispatch<SetStateAction<job[]>>] | null>(null);
