+++ client/src/contexts/sort.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, Dispatch, SetStateAction, useContext, useState } from 'react';
 
-import {sort, Children} from '../../interfaces'
+import {sort, Children} from '../interfaces'
 
 const Context = createContext<[sort, Dispatch<SetStateAction<sort>>] | null>(null);
